PIF interface Creation
We have a PIF interface.
So we are getting close to having the PIF interface working

This is connection to ram internally at the moment. So we can make the PIF rom and ram as one and make the 6502 upload the PIF rom to the N64 PIF ram. then we can make the whole pif ram read and writeable from both the N64 and 6502.