We have the system layout almost completed
I have placed the EEPROM Master core in, but we will need to work on a fifo interface for it.

Also started on the 6502asm code for the cpu to start doing somethings.

Will work on getting it to test the controller and wait for a responce.

After this we can work on the 6105 crc code and the moving of the pif_rom to the pif_ram when booting.

I nice little test on my 6502asm coding ;-)

THere is also work needed to be done on the controller for the mempak interface that needs to be completed.

But for now we can work on this.