Controller init and PIF startup
Here we a little closer. The ASM for the 6502 is coming a long.

Right now I have got most of the core code up and the CRC for the 6105 as well.

There is still more to go and I'm working on the controller right now.

THere is also testing on the verilog code that needs to be done later today to make sure the N64 can send data to the PIF ram and read from it.